Ssim no discriminator

+++ builders/DatasetBuilder.py
@@ -24,7 +24,7 @@ class DatasetBuilder:
     def directory_mapper(self,path):
         img = tf.io.read_file(path)
         img = tf.io.decode_jpeg(img,channels=self.channels)
-        return tf.image.resize(img,[self.height,self.width])
+        return tf.cast(tf.image.resize(img,[self.height,self.width]),tf.uint8)
         
     def build(self):
         if self.is_tfds:
